feat(accord): scalar voting — u64 votes, Median aggregation, bps band
Jurors can now vote a fixed-point scalar (settlement-mint base units,
e.g. 6-dec USDC) instead of only an option index. The Schelling point
for 'how much' disputes is the median of reveals; coherence becomes a
tolerance band rather than exact match (ADR-0025).

- Widen votes/results/rulings u8 -> u64 site-wide (program, SDK, CLI,
  cranker, app, canon-app); u64::MAX sentinels everywhere
- Aggregation::Median — zero-option filing, median tally (even reveal
  count -> upper middle, test-pinned), reveal gate vote != u64::MAX
- coherence_tol_bps on Subaccord/CaseTerms/CreateSubaccordParams
  (default 100 = +/-1% of median, frozen at filing, immutable)
- Commit preimage: hash(vote_le8 || salt || juror) — 72 bytes
- SDK: encodeScalarVote/decodeScalarVote, bigint facades, NO_VOTE/
  NO_RULING sentinels; CLI: --vote decimal-scalar mode + --decimals,
  --aggregation/--coherence-tol-bps, optional --options for scalar
  filing; app: aggregation-aware ruling display + scalar commit form
- Round zero_copy re-tiled to a single 2-byte pad (2824 -> 2816):
  settled/bump fill the u32-block alignment hole
- Fix pre-existing draw.spec fixed-nonce idempotency bug (second suite
  run on one Surfnet collided on the dispute PDA)

All lanes green: Rust 14+63+canon, SDK 97, CLI 119, cranker 56, app 25,
jest e2e 65/65 on Surfpool.

Refs: accord-h5nx, accord-jrgf

+++ programs/accord/src/events.rs
@@ -96,7 +96,7 @@ pub struct Revealed {
     pub dispute: Pubkey,
     pub round_idx: u32,
     pub juror: Pubkey,
-    pub vote: u8,
+    pub vote: u64,
 }
 
 /// Emitted when a round is tallied.
@@ -104,14 +104,14 @@ pub struct Revealed {
 pub struct RoundResolved {
     pub dispute: Pubkey,
     pub round_idx: u32,
-    pub result: u8,
+    pub result: u64,
 }
 
 /// Emitted when a dispute reaches its final ruling (Arbitrable reads this).
 #[event]
 pub struct RulingFinalized {
     pub dispute: Pubkey,
-    pub ruling: u8,
+    pub ruling: u64,
 }
 
 /// Emitted when a prior round's coherence settlement lands via `settle_round`
